requirements.txt
fastapi
uvicorn
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
keras==3.6.0
Markdown==3.7
ml-dtypes==0.4.1
mlflow==2.17.2
mlflow-skinny==2.17.2
numpy==2.0.2
opencv-python==4.10.0.84
opentelemetry-api==1.28.1
opentelemetry-sdk==1.28.1
opentelemetry-semantic-conventions==0.49b1
pandas==2.2.3
parso==0.8.4
pillow==11.0.0
pip-tools==7.4.1
platformdirs==4.3.6
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
Pygments==2.18.0
pyparsing==3.2.0
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
scikit-learn==1.5.2
scipy==1.14.1
six==1.16.0
smmap==5.0.1
SQLAlchemy==2.0.36
sqlparse==0.5.2
stack-data==0.6.3
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.31.0
Werkzeug==3.1.3
wrapt==1.16.0
zipp==3.21.0
